now wait correctly
